// LLD/stock-broker-system/src/enums/Sector.ts
export enum Sector {
  TECHNOLOGY = 'TECHNOLOGY',
  BANKING = 'BANKING',
  FINANCE = 'FINANCE',
  ENERGY = 'ENERGY',
  CONSUMER_GOODS = 'CONSUMER_GOODS',
  AUTOMOBILE = 'AUTOMOBILE',
  PHARMACEUTICAL = 'PHARMACEUTICAL',
  INFRASTRUCTURE = 'INFRASTRUCTURE',
  TELECOM = 'TELECOM',
  METALS = 'METALS',
  REAL_ESTATE = 'REAL_ESTATE',
  FMCG = 'FMCG',
  MEDIA = 'MEDIA',
  HEALTHCARE = 'HEALTHCARE',
  UTILITIES = 'UTILITIES'
}